## What this delivers
  - **Attendees from a dedicated `Zvanica u pratnji` column** (located by header name), replacing the old Napomena name-list heuristic.
  Opt-in: an absent column is non-destructive.
  - **Continue/clean sync modes.** Continue adds/updates (never deletes). Clean mirrors the sheet as one atomic `FOR UPDATE` transaction
  (matched guests keep their id/seat/created_at; orphans deleted) with a typed-DELETE confirmation.
  - **Sheet-row ordering** — `sheet_row` (integer) so the list mirrors the spreadsheet; manual rows sort last.
  - **Multi-tenant invitations controller** — `ClientContextGuard` + `clientId` threaded through every read/write.

  ## Review hardening (last 6 commits)
  - **Critical:** sync no longer wipes all attendees when the sheet lacks the companions column (absent ≠ empty).
  - **High:** clean mode no longer deletes an in-sheet guest over an out-of-range count typo (soft-errored rows are protected from
  orphan-deletion).
  - **High:** dedicated-column decision documented; column-less seed/sync produce zero attendees gracefully.
  - Continue mode is additive (won't delete app-added attendees); clean mirrors.
  - `is_child` no longer mislabels adults when the child count exceeds named companions (TS + Python parity).
  - Clean mode reads `FOR UPDATE` inside the transaction (no TOCTOU / optimistic-lock abort).
  - Clean toast reports `renamed` + `attendeesRemoved`.
  - Cleanups: shared `COUNT_MIN/MAX`, `sheet_row` widened to `integer`, factored normalizers.

  ## Verification
  - API `tsc --noEmit` + `nest build`: pass. Web `tsc -b && vite build`: pass.
  - `npm test`: 13 unit tests pass; the DB integration spec is gated behind `RUN_DB_TESTS=1` (a plain `npm test` skips it, never touches a
  live DB).
  - Integration suite (8 tests, incl. new regressions) verified against a disposable `guests_test` DB; schema + migration 04 validated
  idempotent.

  ## Migration note
  `sheet_row` widened smallint→integer. Apply `db/migrations/04_sheet_row_order.sql` by hand on the live volume (idempotent); back up
  first.

+++ db/01_schema.sql
@@ -119,8 +119,10 @@ CREATE TABLE invitation (
   -- Original Google Sheet row number (1-indexed), captured on sync so the
   -- invitations list can reproduce the spreadsheet's row order instead of
   -- sorting alphabetically. NULL for invitations created manually in the app
-  -- (they sort after all synced rows — see InvitationsService.list).
-  sheet_row        smallint,
+  -- (they sort after all synced rows — see InvitationsService.list). `integer`,
+  -- not smallint: it mirrors an externally-controlled sheet row that can exceed
+  -- the 32767 smallint ceiling.
+  sheet_row        integer,
   created_by       uuid REFERENCES app_user(id) ON DELETE SET NULL,
   updated_by       uuid REFERENCES app_user(id) ON DELETE SET NULL,
   created_at       timestamptz NOT NULL DEFAULT now(),
